Add Rapid Z setting
I found that the retraction speed was too fast for my z-axis. (old shapeko)
It would screech and end up missing the target depth, then crashing through
wood.  What I found needing is a "rapid-z" setting for moving in
the z axis when not cutting through material -- the equivalent of 'rapid'
but not for x/y axis.